Created predicates for find command
