--- src/main/java/seedu/address/model/module/NameContainsKeywordsPredicate.java
@@ -1,31 +1,0 @@
-package seedu.address.model.module;
-
-import java.util.List;
-import java.util.function.Predicate;
-
-import seedu.address.commons.util.StringUtil;
-
-/**
- * Tests that a {@code Module}'s {@code name} matches any of the keywords given.
- */
-public class NameContainsKeywordsPredicate implements Predicate<ReadOnlyModule> {
-    private final List<String> keywords;
-
-    public NameContainsKeywordsPredicate(List<String> keywords) {
-        this.keywords = keywords;
-    }
-
-    @Override
-    public boolean test(ReadOnlyModule module) {
-        return keywords.stream()
-                .anyMatch(keyword -> StringUtil.containsWordIgnoreCase(module.getName().name, keyword));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof NameContainsKeywordsPredicate // instanceof handles nulls
-                && keywords.equals(((NameContainsKeywordsPredicate) other).keywords)); // state check
-    }
-
-}
